Refactor MINIMAL_RUNTIME shell loader
